+++ docs/app/pages/index.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { joinURL } from 'ufo'
 // @ts-expect-error yaml is not typed
-import page from '.index.yml'
+import page from './.index.yml'
 
 const { url } = useSiteConfig()