+++ docs/app/pages/pro/index.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import { joinURL } from 'ufo'
 // @ts-expect-error yaml is not typed
-import page from '.content/pro.yml'
+import page from './.content/pro.yml'
 // @ts-expect-error yaml is not typed
-import templatesPage from '.content/templates.yml'
+import templatesPage from './.content/templates.yml'
 
 const { url } = useSiteConfig()