+++ playground/app/pages/components/file-upload.vue
@@ -46,10 +46,10 @@ const schema = z.object({
             const img = new Image()
             img.onload = () => {
               const meetsDimensions
-                  = img.width >= MIN_DIMENSIONS.width
-                    && img.height >= MIN_DIMENSIONS.height
-                    && img.width <= MAX_DIMENSIONS.width
-                    && img.height <= MAX_DIMENSIONS.height
+                = img.width >= MIN_DIMENSIONS.width
+                  && img.height >= MIN_DIMENSIONS.height
+                  && img.width <= MAX_DIMENSIONS.width
+                  && img.height <= MAX_DIMENSIONS.height
               resolve(meetsDimensions)
             }
             img.src = e.target?.result as string
@@ -62,25 +62,21 @@ const schema = z.object({
     )
 })
 
-type schema = z.output<typeof schema>
+type Schema = z.output<typeof schema>
 
-const state = reactive<Partial<schema>>({
+const state = reactive<Partial<Schema>>({
   avatar: undefined
 })
 
 const value = ref<File | null>(null)
 const valueMultiple = ref<File[]>([])
 
-const upload = useUpload('/api/blob', { method: 'PUT' })
-
 function createObjectUrl(file: File): string {
   return URL.createObjectURL(file)
 }
 
-async function onSubmit(event: FormSubmitEvent<schema>) {
-  const res = await upload(event.data.avatar)
-
-  console.log(res)
+async function onSubmit(event: FormSubmitEvent<Schema>) {
+  console.log(event.data)
 }
 </script>