+++ playground/app/pages/components/modal.vue
@@ -63,7 +63,7 @@ function openFirstModal() {
       <UButton label="Open fullscreen" color="neutral" variant="outline" />
     </UModal>
 
-    <UModal title="Modal prevent close" description="This modal has `dismissible: false` prop so it won't close when clicking outside." :dismissible="false">
+    <UModal title="Modal prevent close" description="This modal has `dismissible: false` prop so it won't close when clicking outside." :dismissible="false" :modal="false" :overlay="false">
       <UButton label="Open unclosable" color="neutral" variant="subtle" />
     </UModal>