+++ src/runtime/components/Carousel.vue
@@ -115,7 +115,7 @@ export interface CarouselEmits {
 </script>
 
 <script setup lang="ts" generic="T extends CarouselItem">
-import { computed, ref, watch, onMounted } from 'vue'
+import { computed, ref, watch, onMounted, onBeforeUnmount } from 'vue'
 import useEmblaCarousel from 'embla-carousel-vue'
 import { Primitive, useForwardProps } from 'reka-ui'
 import { reactivePick } from '@vueuse/core'
@@ -213,13 +213,16 @@ async function loadPlugins() {
   plugins.value = emblaPlugins
 }
 
-watch(() => [props.autoplay, props.autoScroll, props.autoHeight, props.classNames, props.fade, props.wheelGestures], loadPlugins, { immediate: true })
-
-const [emblaRef, emblaApi] = useEmblaCarousel(options.value, plugins.value)
-
-watch([options, plugins], () => {
+watch(() => [props.autoplay, props.autoScroll, props.autoHeight, props.classNames, props.fade, props.wheelGestures], async () => {
+  await loadPlugins()
   emblaApi.value?.reInit(options.value, plugins.value)
-})
+}, { immediate: true })
+
+const [emblaRef, emblaApi] = useEmblaCarousel(options, plugins)
+
+watch(options, () => {
+  emblaApi.value?.reInit(options.value, plugins.value)
+}, { flush: 'post' })
 
 function scrollPrev() {
   emblaApi.value?.scrollPrev()
@@ -232,8 +235,15 @@ function scrollTo(index: number) {
 }
 
 function onKeyDown(event: KeyboardEvent) {
-  const prevKey = props.orientation === 'vertical' ? 'ArrowUp' : 'ArrowLeft'
-  const nextKey = props.orientation === 'vertical' ? 'ArrowDown' : 'ArrowRight'
+  let prevKey
+  let nextKey
+  if (props.orientation === 'horizontal') {
+    prevKey = dir.value === 'ltr' ? 'ArrowLeft' : 'ArrowRight'
+    nextKey = dir.value === 'ltr' ? 'ArrowRight' : 'ArrowLeft'
+  } else {
+    prevKey = 'ArrowUp'
+    nextKey = 'ArrowDown'
+  }
 
   if (event.key === prevKey) {
     event.preventDefault()
@@ -274,11 +284,23 @@ onMounted(() => {
     return
   }
 
-  emblaApi.value?.on('init', onInit)
-  emblaApi.value?.on('init', onSelect)
-  emblaApi.value?.on('reInit', onInit)
-  emblaApi.value?.on('reInit', onSelect)
-  emblaApi.value?.on('select', onSelect)
+  emblaApi.value.on('init', onInit)
+  emblaApi.value.on('init', onSelect)
+  emblaApi.value.on('reInit', onInit)
+  emblaApi.value.on('reInit', onSelect)
+  emblaApi.value.on('select', onSelect)
+})
+
+onBeforeUnmount(() => {
+  if (!emblaApi.value) {
+    return
+  }
+
+  emblaApi.value.off('init', onInit)
+  emblaApi.value.off('init', onSelect)
+  emblaApi.value.off('reInit', onInit)
+  emblaApi.value.off('reInit', onSelect)
+  emblaApi.value.off('select', onSelect)
 })
 
 defineExpose({